requirements.txt
flask
requests
gevent
mysql-connector-python